requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=1.10.0
aiofiles
python-multipart
aiohttp
numpy
librosa
pydub
soundfile
huggingface_hub==0.28.1
tokenizers==0.21.0
transformers==4.49.0
moshi==0.2.2
torchtune==0.4.0
torchao==0.9.0
